Fold the Provider fetch extensions into their use sites
Dissolve Provider+Fetch.swift: the Provider fetch methods move next to
the protocol in Provider.swift, and the Sequence helpers move into
GlobalSearch.swift, their only use site. Result.error tightens to
fileprivate now that its lone client lives in the same file.
