+++ Sources/ScoutUI/Support/Provider/Provider.swift
@@ -18,3 +18,44 @@ protocol Provider: ObservableObject, Periodic {
 
     func fetch(in database: DatabaseReader) async throws -> Output
 }
+
+extension Provider {
+    func fetchAgain(in database: DatabaseReader) async {
+        do {
+            result = .success(try await fetch(in: database))
+        } catch is CancellationError {
+            // A cancelled task (e.g. the view was recreated) leaves the result untouched so it retries.
+        } catch {
+            result = .failure(error)
+        }
+    }
+
+    func fetchIfNeeded(in database: DatabaseReader) async {
+        guard result == nil else {
+            return
+        }
+        await fetchAgain(in: database)
+    }
+
+    func fetchIfFailed(in database: DatabaseReader) async {
+        guard error != nil else {
+            return
+        }
+        await fetchAgain(in: database)
+    }
+
+    @discardableResult
+    func fetchLatest(in database: DatabaseReader) async -> Bool {
+        do {
+            result = .success(try await fetch(in: database))
+            return true
+        } catch is CancellationError {
+            return true
+        } catch {
+            if result == nil {
+                result = .failure(error)
+            }
+            return false
+        }
+    }
+}
